requirements.txt
python-telegram-bot
ollama
requests
ddgs
anthropic
pyyaml
